Separate out TrkInfo structs and add to ntuplehelper
