+++ inc/TrkInfo.hh
@@ -1,59 +1,14 @@
 //
-// structs used to record per-track information in TTrees.
+// TrkInfo: general information about a reconstructed track
 // All momenta are in units of MeV/c, time in nsec WRT when the proton bunch pulse peak hits the production target,
 // positions are in mm WRT the center of the tracker.
 // Dave Brown (LBNL)
 //
 #ifndef TrkInfo_HH
 #define TrkInfo_HH
-#include "TrkAna/inc/RootVectors.hh"
-#include "Offline/DataProducts/inc/SurfaceId.hh"
+
 namespace mu2e
 {
-  // generic information about the track fit at a particular place/time
-  struct TrkFitInfo {
-    XYZVectorF mom;
-    XYZVectorF pos;
-    double time;
-    float momerr = -1000;  // projected error on the scalar momentum
-    bool inbounds = false; // was the intersection within the (literal) surface bounds
-    bool gap = false; // was the intersection in a piecewise trajectory gap
-    bool early = false; // is this the earliest intersection for this track
-    bool late = false; // is this the latest intersection for this track
-    int sid = SurfaceIdEnum::unknown; // SurfaceId of the intersected surface, see Offline/KinKalGeom/inc/SurfaceId.hh
-    int sindex = 0; // index to the intersected surface (for multi-surface elements like StoppingTarget)
-    void reset() { *this = TrkFitInfo(); }
-  };
-  // structs for specific fit types
-  struct CentralHelixInfo {
-    // central helix parameters
-    float d0=0,phi0=0,omega=0,z0=0,tanDip=0,t0=0;
-    float d0err=0,phi0err=0,omegaerr=0,z0err=0,tanDiperr=0,t0err=0;
-    // max radius (assuming geometric extrapolation).  This member is deprecated
-    // in favor of explicit extrapolation to the OPA
-    float maxr = 0;
-    void reset() { *this = CentralHelixInfo(); }
-  };
-
-  struct LoopHelixInfo {
-    //  These members are deprecated
-    // in favor of explicit extrapolation to the OPA
-    float maxr = 0; // max radius (assuming geometric extrapolation).
-    float d0 = 0; // d0 (unsigned!)
-    float tanDip = 0; // tanDip
-    // loop helix parameters
-    float rad=0,lam=0,cx=0,cy=0,phi0=0,t0=0;
-    float raderr=0,lamerr=0,cxerr=0,cyerr=0,phi0err=0,t0err=0;
-    void reset() { *this = LoopHelixInfo(); }
-  };
-
-  struct KinematicLineInfo {
-    float  d0=0, phi0=0, z0=0, theta=0, mom=0,t0=0;
-    float  d0err=0, phi0err=0, z0err=0, thetaerr=0, momerr=0,t0err=0;
-    void reset() { *this = KinematicLineInfo(); }
-  };
-
-  // general information about a track
   struct TrkInfo {
     int status =-1; // Kalman fit status
     int goodfit = 0; // fit is geometrically consistent
@@ -81,18 +36,6 @@ namespace mu2e
     float maxgap =-1, avggap =-1; // fit trajectory gaps
     float avgedep = 0; // average active hit edep;
     void reset() { *this = TrkInfo(); }
-  };
-
-  // general NC info about the particle which generated the majority of the hits used in this track
-  struct TrkInfoMC {
-    bool valid =false; // whether this information is valid or not
-    int ndigi =0; // number of true straw digitizations (both ends) generatedy by the primary particle
-    int  ndigigood =0; // number of true straw digitizations where the primary particle (not a delta-ray) crossed threshold
-    int nhits =0, nactive =0; // number of hits on the track and active on the track generated by the primary particle
-    int nambig =0; // number of true hits where the reconstruction assigned the correct left-right ambiguity
-    float maxr = 0.; // max radius calculated using the LoopHelix constructor
-    float rad=0,lam=0,cx=0,cy=0,phi0=0,t0=0; // truth track parameters
-    void reset() { *this = TrkInfoMC(); }
   };
 }
 #endif
