+++ inc/TrkInfoMC.hh
@@ -1,0 +1,23 @@
+//
+// TrkInfoMC: general MC info about the particle which generated the majority of the hits used in this track
+// All momenta are in units of MeV/c, time in nsec WRT when the proton bunch pulse peak hits the production target,
+// positions are in mm WRT the center of the tracker.
+// Dave Brown (LBNL)
+//
+#ifndef TrkInfoMC_HH
+#define TrkInfoMC_HH
+
+namespace mu2e
+{
+  struct TrkInfoMC {
+    bool valid =false; // whether this information is valid or not
+    int ndigi =0; // number of true straw digitizations (both ends) generatedy by the primary particle
+    int  ndigigood =0; // number of true straw digitizations where the primary particle (not a delta-ray) crossed threshold
+    int nhits =0, nactive =0; // number of hits on the track and active on the track generated by the primary particle
+    int nambig =0; // number of true hits where the reconstruction assigned the correct left-right ambiguity
+    float maxr = 0.; // max radius calculated using the LoopHelix constructor
+    float rad=0,lam=0,cx=0,cy=0,phi0=0,t0=0; // truth track parameters
+    void reset() { *this = TrkInfoMC(); }
+  };
+}
+#endif
