+++ src/TrkAnaTreeMaker_module.cc
@@ -68,6 +68,11 @@
 #include "TrkAna/inc/EventInfo.hh"
 #include "TrkAna/inc/EventInfoMC.hh"
 #include "TrkAna/inc/TrkInfo.hh"
+#include "TrkAna/inc/TrkInfoMC.hh"
+#include "TrkAna/inc/TrkFitInfo.hh"
+#include "TrkAna/inc/LoopHelixInfo.hh"
+#include "TrkAna/inc/CentralHelixInfo.hh"
+#include "TrkAna/inc/KinematicLineInfo.hh"
 #include "TrkAna/inc/SimInfo.hh"
 #include "TrkAna/inc/EventWeightInfo.hh"
 #include "TrkAna/inc/TrkStrawHitInfo.hh"
