+++ validation/ntuplehelper-test.py
@@ -4,7 +4,7 @@ nthelp = ntuplehelper.nthelp()
 
 print("Testing a leaf from each branch")
 print("=========")
-nthelp.whatis(['evtinfomc.pbtime', 'evtinfo.event', 'hcnt.nbkg', 'tcnt.ndem', 'demmcsim.pdg', 'demtch.trkdepth', 'demtsh.ptoca', 'demtshmc.edep', 'demtsm.tlen', 'crvsummary.totalPEs', 'crvsummarymc.sectorNumber', 'crvcoincs.time', 'crvcoincsmc.primaryE', 'crvcoincsmcplane.dir', 'demmcsci.dp', 'demmcssi.dp'])
+nthelp.whatis(['evtinfomc.pbtime', 'evtinfo.event', 'hcnt.nbkg', 'tcnt.ndem', 'demmcsim.pdg', 'demtch.trkdepth', 'demtsh.ptoca', 'demtshmc.edep', 'demtsm.tlen', 'crvsummary.totalPEs', 'crvsummarymc.sectorNumber', 'crvcoincs.time', 'crvcoincsmc.primaryE', 'crvcoincsmcplane.dir', 'demmcsci.dp', 'demmcssi.dp', 'dem.nactive', 'demmc.ndigigood', 'demfit.mom', 'demlh.cx', 'demch.omega', 'klkl.phi0'])
 
 print("Testing a few leaves from a single branch (should only print the branch description once)")
 print("=========")
